requirements.txt
PyMuPDF==1.22.5
pdf2image==1.16.3
pytesseract==0.3.10
deepl==1.15.0
Flask==2.3.3
Werkzeug==2.3.7
python-dotenv==1.0.0
Pillow==10.0.0 